requirements.txt
streamlit
opencv-python-headless
mediapipe==0.10.9
ffmpeg-python
numpy
protobuf==3.20.3
